buttons show up
