+++ src/SceneGame.js
@@ -1,3 +1,4 @@
+import ArmyInfoScene from "./ui/scenes/ArmyInfoScene";
 
 import GameUtils from "./utils/GameUtils";
 import GameUtilsUi from "./utils/GameUtilsUi";
@@ -6,8 +7,6 @@ import GameUtilsBuilding from "./utils/GameUtilsBuilding";
 import Board from "./board/Board";
 import TerrainObj from "./board/Terrain";
 const { Terrain, TerrainSpriteName } = TerrainObj;
-import DirectionObj from "./board/Direction";
-const { Direction } = DirectionObj;
 
 import Village from "./buildings/villageBuildings/Village";
 
@@ -452,151 +451,6 @@ export default class SceneGame extends Phaser.Scene {
         this.uiBuilding.push(this.txtBuildName);
         this.uiBuilding.push(this.btnBuildDestroy);
 
-        //TODO: buttons have all clickedFunctions 
-        /**
-         * UI - army
-         */
-        //TODO: pull this out to a scene on top of another scene.
-
-        x = -375;
-        y = -120;
-
-        this.txtArmySize = this.add.text(x, y)
-            .setScrollFactor(0)
-            .setFontSize(50)
-            .setDepth(100)
-            .setShadow(3, 3, "#000000", 3);
-
-        this.txtArmyVillage = this.add.text(x, y + 60)
-            .setScrollFactor(0)
-            .setFontSize(50)
-            .setDepth(100)
-            .setShadow(3, 3, "#000000", 3);
-
-        this.txtArmyMoves = this.add.text(x, y + 120)
-            .setScrollFactor(0)
-            .setFontSize(50)
-            .setDepth(100)
-            .setShadow(3, 3, "#000000", 3);
-
-        this.txtArmyFood = this.add.text(x, y + 180)
-            .setScrollFactor(0)
-            .setFontSize(50)
-            .setDepth(100)
-            .setShadow(3, 3, "#000000", 3);
-
-        this.txtArmyWood = this.add.text(x, y + 240)
-            .setScrollFactor(0)
-            .setFontSize(50)
-            .setDepth(100)
-            .setShadow(3, 3, "#000000", 3);
-
-        this.btnArmyGetUnits = this.add.sprite(x, y + 290, "btnArmyGetUnits")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setDepth(100)
-            .setOrigin(0)
-            .on("pointerdown", this.armyManager.armyGetUnits);
-
-        this.btnArmyDisbandUnits = this.add.sprite(x, y + 430, "btnArmyDisbandUnits")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setDepth(100)
-            .setOrigin(0)
-            .on("pointerdown", this.armyManager.armyDisbandUnits);
-
-        this.btnArmyGetFood = this.add.sprite(x, y + 570, "btnArmyGetFood")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setDepth(100)
-            .setOrigin(0)
-            .on("pointerdown", this.armyManager.armyGetFood);
-
-        this.btnArmyGetWood = this.add.sprite(x, y + 710, "btnArmyGetWood")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setDepth(100)
-            .setOrigin(0)
-            .on("pointerdown", this.armyManager.armyGetWood);
-
-        this.btnArmyBuild = this.add.sprite(x, y + 850, "btnArmyBuild")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setDepth(100)
-            .setOrigin(0)
-            .on("pointerdown", this.armyManager.armyBuild);
-        //TODO: clicked
-
-        //TODO: alot of repeated code here and .setscrollfactor, etc
-        this.uiArmyText.push(this.txtArmySize);
-        this.uiArmyText.push(this.txtArmyVillage);
-        this.uiArmyText.push(this.txtArmyMoves);
-        this.uiArmyText.push(this.txtArmyFood);
-        this.uiArmyText.push(this.txtArmyWood);
-        this.uiArmyButtons.push(this.btnArmyGetUnits);
-        this.uiArmyButtons.push(this.btnArmyDisbandUnits);
-        this.uiArmyButtons.push(this.btnArmyGetFood);
-        this.uiArmyButtons.push(this.btnArmyGetWood);
-        this.uiArmyButtons.push(this.btnArmyBuild);
-
-        this.btnArmyBuildCancel = this.add.sprite(x, y + 290, "btnArmyCancel")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setOrigin(0)
-            .setDepth(100)
-            .on("pointerdown", this.armyManager.armyBuildCancel);
-
-        this.btnArmyBuildWallWood = this.add.sprite(x, y + 430, "btnArmyBuildWallWood")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setDepth(100)
-            .setOrigin(0)
-            .on("pointerdown", this.clickedBuildWallWood);
-
-        this.uiArmyBuildButtons.push(this.btnArmyBuildCancel);
-        this.uiArmyBuildButtons.push(this.btnArmyBuildWallWood);
-
-        this.btnArmyBuildEast = this.add.sprite(x + 420, y + 360, "btnArmyBuildEast")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setOrigin(0)
-            .setDepth(100)
-            .on("pointerdown", function () {
-                this.scene.selectedArmyBuildFunc(Direction.EAST);
-            });
-
-        this.btnArmyBuildNorth = this.add.sprite(x + 350, y + 290, "btnArmyBuildNorth")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setOrigin(0)
-            .setDepth(100)
-            .on("pointerdown", function () {
-                this.scene.selectedArmyBuildFunc(Direction.NORTH);
-            });
-
-        this.btnArmyBuildSouth = this.add.sprite(x + 350, y + 430, "btnArmyBuildSouth")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setOrigin(0)
-            .setDepth(100)
-            .on("pointerdown", function () {
-                this.scene.selectedArmyBuildFunc(Direction.SOUTH);
-            });
-
-        this.btnArmyBuildWest = this.add.sprite(x + 280, y + 360, "btnArmyBuildWest")
-            .setScrollFactor(0)
-            .setInteractive()
-            .setOrigin(0)
-            .setDepth(100)
-            .on("pointerdown", function () {
-                this.scene.selectedArmyBuildFunc(Direction.WEST);
-            });
-
-        this.uiArmyBuildButtons.push(this.btnArmyBuildEast);
-        this.uiArmyBuildButtons.push(this.btnArmyBuildNorth);
-        this.uiArmyBuildButtons.push(this.btnArmyBuildSouth);
-        this.uiArmyBuildButtons.push(this.btnArmyBuildWest);
-
         /**
          * ui enemy elements
          */
@@ -702,8 +556,8 @@ export default class SceneGame extends Phaser.Scene {
             up: keys.W,
             down: keys.S,
             //TODO: this causes UI to shrink/expand. i dont know how to fix this now
-            //zoomIn: this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.Q),
-            //zoomOut: this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.E),
+            zoomIn: this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.Q),
+            zoomOut: this.input.keyboard.addKey(Phaser.Input.Keyboard.KeyCodes.E),
             acceleration: 1.00,
             drag: 0.01,
             maxSpeed: 1.0
@@ -763,6 +617,13 @@ export default class SceneGame extends Phaser.Scene {
 
         //TODO:center to player 1 center. remove? make more dynamic?
         this.cam.pan(1000, 2000, 1000);
+
+        var handle = "armyinfoscene";
+        var demo = new ArmyInfoScene(this);
+
+        //this.input.setDraggable(win);
+
+        this.scene.add(handle, demo, true);
     }
 
     update(time, delta) {
@@ -924,6 +785,7 @@ export default class SceneGame extends Phaser.Scene {
         console.log("ending turn: player2...");
     }
 
+    //TODO: refactor to GameEngine class
     //replenishment
     preTurnPhase(playerNumber) {
 
@@ -1061,9 +923,6 @@ export default class SceneGame extends Phaser.Scene {
     deselectEverything() {
         GameUtilsUi.hideGameObjects(this.uiVillage);
         GameUtilsUi.hideGameObjects(this.uiBuilding);
-        GameUtilsUi.hideGameObjects(this.uiArmyText);
-        GameUtilsUi.hideGameObjects(this.uiArmyButtons);
-        GameUtilsUi.hideGameObjects(this.uiArmyBuildButtons);
         GameUtilsUi.hideGameObjects(this.uiArmyEnemy);
         GameUtilsUi.hideGameObjects(this.uiArmyEnemyBuilding);
 
@@ -1210,89 +1069,6 @@ export default class SceneGame extends Phaser.Scene {
 
         }
 
-    }
-
-    //TODO: refactor elsewhere
-
-    /**
-     * shows the ui army main action buttons
-     * hides non-main action buttons
-     * @param {Army} armyData 
-     */
-    showUiArmyButtons(armyData) {
-        //TODO: refactor armyData as just selected
-        GameUtilsUi.hideGameObjects(this.uiArmyBuildButtons);
-        this.selectedArmyBuildFunc = null;
-
-        let row = armyData.row;
-        let col = armyData.col;
-
-        let buildingSprite = this.board.boardBuildings[row][col];
-
-        //on-top-of-village buttons
-        //TODO: be able to replenish in friendly villages through trade.
-        //TODO: buttons in a list. text in a list.
-        this.btnArmyGetFood.visible = false;
-        this.btnArmyGetUnits.visible = false;
-        this.btnArmyDisbandUnits.visible = false;
-        if (buildingSprite != null) {
-
-            let buildingData = buildingSprite.data.get("data");
-
-            /** @type {Village} */
-            let village = buildingData.village;
-
-            if (village == null)
-                return;
-
-            //if this is your territory
-            if (buildingData.player == armyData.player) {
-                GameUtilsUi.showGameObjects(this.uiArmyButtons);
-                GameUtils.clearTintArray(this.uiArmyButtons);
-
-                /**
-                 * adequate resources check
-                 */
-                //get food
-                if (village.amountFood < 10)
-                    this.btnArmyGetFood.setTint("0xff0000");
-                else if (village.amountFood == 10)
-                    this.btnArmyGetFood.setTint("0xffff00");
-
-                //get units
-                if (village.population < 10)
-                    this.btnArmyGetUnits.setTint("0xff0000");
-                else if (village.population == 10)
-                    this.btnArmyGetUnits.setTint("0xffff00");
-
-                //get wood
-                if (village.amountWood < 10)
-                    this.btnArmyGetWood.setTint("0xff0000");
-                else if (village.amountWood == 10)
-                    this.btnArmyGetWood.setTint("0xffff00");
-            }
-        }
-
-    }
-
-    //TODO: refactor elsewhere
-    /**
-     * shows the army build buttons.
-     * hides all non army build buttons
-     */
-    showUiArmyBuildButtons() {
-        GameUtilsUi.hideGameObjects(this.uiArmyButtons);
-        GameUtilsUi.showGameObjects(this.uiArmyBuildButtons);
-        GameUtils.clearTintArray(this.uiArmyBuildButtons);
-
-        /** @type {Army} */
-        let armyData = this.selectedArmy.getData("data");
-
-        //TODO: remove hardcode. maybe to json file -> load singleton
-        if (armyData.amountWood < 100)
-            this.btnArmyBuildWallWood.setTint("0xff0000");
-
-        this.updateUiArmyBuildDirectionButtons();
     }
 
     updateUiArmyBuildDirectionButtons() {
